Update index.js
In this I remove unnecessary lines, resulting in faster installations and have made significant improvements to comments, meaning that users will be more informed whilst using the package and making it more understandable. Additionally, in the `get()` method, if it fails to parse the value as JSON, it will return the raw `strValue` instead, making this package more robust and less of a pain to use.